More size_t

+++ src/Backends/Misc.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stddef.h>
 #include <string>
 
 #include "../Attributes.h"
@@ -90,8 +91,8 @@ void Backend_Deinit(void);
 void Backend_PostWindowCreation(void);
 bool Backend_GetBasePath(std::string *string_buffer);
 void Backend_HideMouse(void);
-void Backend_SetWindowIcon(const unsigned char *rgb_pixels, unsigned int width, unsigned int height);
-void Backend_SetCursor(const unsigned char *rgb_pixels, unsigned int width, unsigned int height);
+void Backend_SetWindowIcon(const unsigned char *rgb_pixels, size_t width, size_t height);
+void Backend_SetCursor(const unsigned char *rgb_pixels, size_t width, size_t height);
 void PlaybackBackend_EnableDragAndDrop(void);
 bool Backend_SystemTask(bool active);
 void Backend_GetKeyboardState(bool *keyboard_state);
